Em-82 prettier export functions
