+++ src/Components/Forms/TestsAndTreatments/Helpers/DataHelpers.js
@@ -10,63 +10,9 @@ import FormattedInputs from 'Components/Generic/MaskedControllers/FormattedInput
 import {
   handleVarientCustomClickFunction,
   handleVarientPaddTheZeroPlaceClickFunction,
+  mergeMultipleIndicators,
+  thickenWithDataFunction,
 } from 'Components/Forms/TestsAndTreatments/Helpers/FunctionHelpers';
-import { thickenWithDataFunction } from './FunctionHelpers';
-
-function mergeMultipleConstants(
-  variantIndicatorsNormalizedData,
-  keyOne,
-  keyTwo,
-  seperator,
-) {
-  let keysPlaces = [];
-  Object.entries(variantIndicatorsNormalizedData).map(([key, dataset]) => {
-    if (
-      dataset &&
-      (dataset['description'] === keyOne || dataset['description'] === keyTwo)
-    ) {
-      keysPlaces.push(key);
-    }
-  });
-  if (keysPlaces.length > 0) {
-    let variantIndicatorsNormalizedDataTemp = JSON.parse(
-      JSON.stringify(variantIndicatorsNormalizedData),
-    );
-    variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['description'] = `${
-      variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['description']
-    }${seperator}${
-      variantIndicatorsNormalizedDataTemp[keysPlaces[1]]['description']
-    }`;
-    variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['unit'] = `${
-      variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['unit']
-    }${seperator}${variantIndicatorsNormalizedDataTemp[keysPlaces[1]]['unit']}`;
-    if (
-      variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['value'] &&
-      variantIndicatorsNormalizedDataTemp[keysPlaces[1]]['value']
-    ) {
-      variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['value'] = `${
-        variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['value']
-          ? variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['value']
-          : ''
-      }${seperator}${
-        variantIndicatorsNormalizedDataTemp[keysPlaces[1]]['value']
-          ? variantIndicatorsNormalizedDataTemp[keysPlaces[1]]['value']
-          : ''
-      }`;
-    }
-    variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['mask'] = `${
-      variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['mask']
-    }${seperator}${variantIndicatorsNormalizedDataTemp[keysPlaces[1]]['mask']}`;
-    variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['code'] = `${
-      variantIndicatorsNormalizedDataTemp[keysPlaces[0]]['code']
-    }${seperator}${variantIndicatorsNormalizedDataTemp[keysPlaces[1]]['code']}`;
-
-    delete variantIndicatorsNormalizedDataTemp[keysPlaces[1]];
-
-    return variantIndicatorsNormalizedDataTemp;
-  }
-  return variantIndicatorsNormalizedData;
-}
 
 export const thickenTheConstantIndicators = ({
   normalizedConstantObservation,
@@ -139,7 +85,7 @@ export const thickenTheVariantIndicators = ({
       : variantIndicatorsNormalizedDataTemp;
   sizeTemp++;
 
-  variantIndicatorsNormalizedDataTemp = mergeMultipleConstants(
+  variantIndicatorsNormalizedDataTemp = mergeMultipleIndicators(
     variantIndicatorsNormalizedDataTemp,
     'Diastolic blood pressure',
     'Systolic blood pressure',
